Produce FeatureBuildItem from extensions even when they are disabled
