+++ extensions/oidc/deployment/src/main/java/io/quarkus/oidc/deployment/OidcAlwaysEnabledProcessor.java
@@ -1,0 +1,15 @@
+package io.quarkus.oidc.deployment;
+
+import io.quarkus.deployment.Feature;
+import io.quarkus.deployment.annotations.BuildStep;
+import io.quarkus.deployment.builditem.FeatureBuildItem;
+
+// Executed even if the extension is disabled, see https://github.com/quarkusio/quarkus/pull/26966/
+public class OidcAlwaysEnabledProcessor {
+
+    @BuildStep
+    FeatureBuildItem featureBuildItem() {
+        return new FeatureBuildItem(Feature.OIDC);
+    }
+
+}
