+++ extensions/opentelemetry/opentelemetry-exporter-otlp/deployment/src/main/java/io/quarkus/opentelemetry/exporter/otlp/deployment/OtlpExporterAlwaysEnabledProcessor.java
@@ -1,0 +1,15 @@
+package io.quarkus.opentelemetry.exporter.otlp.deployment;
+
+import io.quarkus.deployment.Feature;
+import io.quarkus.deployment.annotations.BuildStep;
+import io.quarkus.deployment.builditem.FeatureBuildItem;
+
+// Executed even if the extension is disabled, see https://github.com/quarkusio/quarkus/pull/26966/
+public class OtlpExporterAlwaysEnabledProcessor {
+
+    @BuildStep
+    FeatureBuildItem feature() {
+        return new FeatureBuildItem(Feature.OPENTELEMETRY_OTLP_EXPORTER);
+    }
+
+}
