+++ extensions/opentelemetry/opentelemetry-exporter-otlp/deployment/src/main/java/io/quarkus/opentelemetry/exporter/otlp/deployment/OtlpExporterProcessor.java
@@ -3,12 +3,10 @@ package io.quarkus.opentelemetry.exporter.otlp.deployment;
 import java.util.function.BooleanSupplier;
 
 import io.quarkus.arc.deployment.AdditionalBeanBuildItem;
-import io.quarkus.deployment.Feature;
 import io.quarkus.deployment.annotations.BuildStep;
 import io.quarkus.deployment.annotations.BuildSteps;
 import io.quarkus.deployment.annotations.ExecutionTime;
 import io.quarkus.deployment.annotations.Record;
-import io.quarkus.deployment.builditem.FeatureBuildItem;
 import io.quarkus.deployment.builditem.LaunchModeBuildItem;
 import io.quarkus.opentelemetry.exporter.otlp.runtime.OtlpExporterConfig;
 import io.quarkus.opentelemetry.exporter.otlp.runtime.OtlpExporterProvider;
@@ -23,11 +21,6 @@ public class OtlpExporterProcessor {
         public boolean getAsBoolean() {
             return otlpExporterConfig.enabled;
         }
-    }
-
-    @BuildStep
-    FeatureBuildItem feature() {
-        return new FeatureBuildItem(Feature.OPENTELEMETRY_OTLP_EXPORTER);
     }
 
     @BuildStep
