+++ extensions/security-webauthn/deployment/src/main/java/io/quarkus/security/webauthn/deployment/QuarkusSecurityWebAuthnFeatureAlwaysEnabledProcessor.java
@@ -1,0 +1,15 @@
+package io.quarkus.security.webauthn.deployment;
+
+import io.quarkus.deployment.Feature;
+import io.quarkus.deployment.annotations.BuildStep;
+import io.quarkus.deployment.builditem.FeatureBuildItem;
+
+// Executed even if the extension is disabled, see https://github.com/quarkusio/quarkus/pull/26966/
+class QuarkusSecurityWebAuthnFeatureAlwaysEnabledProcessor {
+
+    @BuildStep
+    FeatureBuildItem feature() {
+        return new FeatureBuildItem(Feature.SECURITY_WEBAUTHN);
+    }
+
+}
